[2.x] Defer to Fortify for password confirmation (#835)
* Consistent handling of password confirmation

* Syntax issues

* Style issues

* Style issues

* Style issues

+++ src/Http/Controllers/Inertia/CurrentUserController.php
@@ -8,6 +8,7 @@ use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
 use Inertia\Inertia;
+use Laravel\Fortify\Actions\ConfirmPassword;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class CurrentUserController extends Controller
@@ -16,18 +17,24 @@ class CurrentUserController extends Controller
      * Delete the current user.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \Illuminate\Contracts\Auth\StatefulGuard  $auth
+     * @param  \Illuminate\Contracts\Auth\StatefulGuard  $guard
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request, StatefulGuard $auth)
+    public function destroy(Request $request, StatefulGuard $guard)
     {
-        $request->validate([
-            'password' => 'required|string|password',
-        ]);
+        $confirmed = app(ConfirmPassword::class)(
+            $guard, $request->user(), $request->password
+        );
+
+        if (! $confirmed) {
+            throw ValidationException::withMessages([
+                'password' => 'The password is incorrect.',
+            ]);
+        }
 
         app(DeletesUsers::class)->delete($request->user()->fresh());
 
-        $auth->logout();
+        $guard->logout();
 
         $request->session()->invalidate();
         $request->session()->regenerateToken();
